move getDialogue functions into more appropriate location
