+++ packages/lesswrong/server/posts/utils.ts
@@ -1,4 +1,5 @@
 import Localgroups from '../../lib/collections/localgroups/collection';
+import { cheerioParse } from "../utils/htmlUtil";
 
 export async function getDefaultPostLocationFields(post: DbPost) {
   if (post.isEvent && post.groupId && !post.location) {
@@ -8,4 +9,25 @@ export async function getDefaultPostLocationFields(post: DbPost) {
     return { location, googleLocation, mongoLocation }
   }
   return {}
+}
+
+export const getDialogueResponseCount = (post:DbPost) => {
+  const html = post.contents.originalContents.data;
+  const parsedHtml= cheerioParse(html);
+  
+  const blocksWithId = parsedHtml('block[message-id]');
+  const ids : string[] = blocksWithId.map( (i, block) => parsedHtml(block).attr('message-id')).get();
+  
+  return ids.length;
+}
+
+export const getDialogueMessageTimestamps = (post: DbPost): Date[] => {
+  const html = post.contents.originalContents.data;
+  const parsedHtml= cheerioParse(html);
+  
+  const blocksWithId = parsedHtml('block[message-id]');
+  const timestampStrings = blocksWithId.map( (i, block) => (parsedHtml(block).attr('submitted-at'))).get();
+  const timestamps = timestampStrings.map( dateString => new Date(dateString))
+  
+  return timestamps
 }
